requirements.txt
langchain
langchain_community 
langgraph
langchain-groq
google-generativeai
langchain-google-genai
python-dotenv
pandas
openpyxl
streamlit
# -e .